VACMS-21017: Moves news release to products
