+++ .github/CODEOWNERS
@@ -17,3 +17,4 @@ src/products/eventListing/ @department-of-veterans-affairs/ap-team
 src/products/staffProfile/ @department-of-veterans-affairs/ap-team
 src/products/newsStory/ @department-of-veterans-affairs/ap-team
 src/products/storyListing/ @department-of-veterans-affairs/ap-team
+src/products/pressRelease/ @department-of-veterans-affairs/ap-team
