+++ src/data/queries/index.ts
@@ -29,7 +29,7 @@ import * as NewsStoryTeaser from '@/data/queries/newsStoryTeaser'
 import * as NumberCallout from '@/data/queries/numberCallout'
 import * as PersonProfile from '@/products/staffProfile/query'
 import * as PhoneNumber from '@/data/queries/phoneNumber'
-import * as PressRelease from '@/data/queries/pressRelease'
+import * as PressRelease from '@/products/pressRelease/query'
 import * as PressReleaseListing from '@/data/queries/pressReleaseListing'
 import * as PressReleaseTeaser from '@/data/queries/pressReleaseTeaser'
 import * as ProcessList from '@/data/queries/processList'
