+++ src/data/queries/tests/pressReleaseTeaser.test.tsx
@@ -1,6 +1,6 @@
 import { NodePressRelease } from '@/types/drupal/node'
 import { queries } from '@/data/queries'
-import mockData from '@/mocks/pressRelease.mock.json'
+import mockData from '@/products/pressRelease/mock.json'
 import { RESOURCE_TYPES } from '@/lib/constants/resourceTypes'
 import { params, formatter } from '../pressReleaseTeaser'
 import { de } from '@faker-js/faker'
