+++ src/pages/[[...slug]].tsx
@@ -39,7 +39,7 @@ import { Event as FormattedEvent } from '@/products/event/formatted-type'
 import { EventListing as FormattedEventListing } from '@/products/eventListing/formatted-type'
 import { LocationsListing as FormattedLocationsListing } from '@/types/formatted/locationsListing'
 import { NewsStory as FormattedNewsStory } from '@/products/newsStory/formatted-type'
-import { PressRelease as FormattedPressRelease } from '@/types/formatted/pressRelease'
+import { PressRelease as FormattedPressRelease } from '@/products/pressRelease/formatted-type'
 import { PressReleaseListing as FormattedPressReleaseListing } from '@/types/formatted/pressReleaseListing'
 import { ResourcesSupport as FormattedResourcesSupport } from '@/types/formatted/resourcesSupport'
 import { StaffProfile as FormattedStaffProfile } from '@/products/staffProfile/formatted-type'
@@ -57,7 +57,7 @@ import { LayoutProps } from '@/templates/layouts/wrapper'
 import { LocationsListing } from '@/templates/layouts/locationsListing'
 import { Meta } from '@/templates/common/meta'
 import { NewsStory } from '@/products/newsStory/template'
-import { PressRelease } from '@/templates/layouts/pressRelease'
+import { PressRelease } from '@/products/pressRelease/template'
 import { PressReleaseListing } from '@/templates/layouts/pressReleaseListing'
 import { PreviewCrumb } from '@/templates/common/preview'
 import { ResourcesSupport } from '@/templates/layouts/resourcesSupport'
