+++ src/products/pressRelease/template.tsx
@@ -1,4 +1,4 @@
-import { PressRelease as FormattedPressRelease } from '@/types/formatted/pressRelease'
+import { PressRelease as FormattedPressRelease } from '@/products/pressRelease/formatted-type'
 import { LovellStaticPropsResource } from '@/lib/drupal/lovell/types'
 import { LovellSwitcher } from '@/templates/components/lovellSwitcher'
 import { formatDate } from '@/lib/utils/helpers'
